requirements.txt
pandas==2.2.3
kafka-python==2.0.2
pyspark==3.5.1
pyarrow==15.0.2
six==1.16.0
